Add support for configurable warnings
New CLI option --warn-conf accepts a comma-separated sequence of
<filter>:<action> pairs for customizing warnings.

Supported filters are:
* any         - matches all warnings
* msg=<regex> - matches warnings where <regex> matches the warning text
* src=<regex> - matches warnings where <regex> matches the warning
                source locator

msg regex is a partial match, it implicitly has '.*' attached to the
beginning and end.

src regex matches any path segment, this is implemented by implicitly
adding (.*/)? to the beginning of the regex.

msg and src filters can be combined with &.

Supported actions are:
* :s - suppress the warning
* :w - report the warning as a warning (default behavior)
* :e - error on the warning

For example, the new Vec dynamic width matching warning can be
suppressed while all other warnings elevated to errors with:
  --warn-conf "msg=Dynamic index.*too (wide|narrow):s,any:e"